Make ProtocolStatus and ParseStatus a VersionedWritable - this helps to manage different versions of segment data.
Add lastModified field to ProtocolStatus - this is useful to keep accessible
in order to quickly calculate fetch times.

Correct naming of some status codes, so that they are more consistent.

NOTE: Segments created with the code base _later_ than rev. r179436
(2005-06-02) will have to be re-created, because there is no way to keep
compatibility with their format.


git-svn-id: https://svn.eu.apache.org/repos/asf/lucene/nutch/trunk@189747 13f79535-47bb-0310-9956-ffa450edef68