Fix two problems:
  1)  A task was reporting umbilical.done() even after an
exception was thrown.  That's wrong; umbilical.done() indicates
the task completed successfully.  We now move the call to umbilical.done()
before the exception-handling

  2)  Completed map tasks sometimes had final progress reports that were
vanishingly close to 1.0, but not quite there.  We have a guard in place
that prevented progress reports that are "too small to matter", and the
final progress report would sometimes fall into that category.  We now
remove the guard when progress reports are within epsilon of 1.0.




git-svn-id: https://svn.eu.apache.org/repos/asf/lucene/nutch/branches/mapred@225387 13f79535-47bb-0310-9956-ffa450edef68