Fix 2 bugs:
  1) Zero-length NDFSOutputStream bug that was introduced as part 
of a buffer-fix last night.  Now we test to see if we're closing down
the stream.  If so, it's OK to send a zero.  Otherwise, not.

  2) File namespace deletions were broken, reporting success but
not actually performing the file delete.  This was introduced as
part of the namespace-handler rewrite earlier today.  (I'm so tired,
I just wrote "a couple of days ago" by mistake.)



git-svn-id: https://svn.eu.apache.org/repos/asf/lucene/nutch/branches/mapred@226950 13f79535-47bb-0310-9956-ffa450edef68